+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 1. [](#improved)
     * Improved error detection for broken Flex Objects
     * Removed apc and xcache support, made apc alias of apcu
+    * Fixed non-namespaced exceptions in scheduler
 
 # v1.6.0-rc.1
 ## 01/30/2019
@@ -19,7 +20,6 @@
     * Fixed `FlexMediaTrait::getMedia()` trying to include uploaded but already moved media
     * Fixed `ImageMedium` constructor warning when file does not exist
     * Fixed bad host header in PSR-7 (if using 'php -S localhost:8000 system/router.php')
-    * Fixed non-namespaced exceptions in scheduler
     
 # v1.6.0-beta.8
 ## 01/25/2019